requirements.txt
mem0ai==0.1.0
chromadb==0.4.15
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6